ParserContext rework and new tests
